move config hashing to runner
