+++ packages/wotan/src/runner.ts
@@ -9,11 +9,13 @@ import {
     MessageHandler,
     FileFilterFactory,
     Severity,
+    ReducedConfiguration,
+    EffectiveConfiguration,
 } from '@fimbul/ymir';
 import * as path from 'path';
 import * as ts from 'typescript';
 import * as glob from 'glob';
-import { unixifyPath, hasSupportedExtension, addUnique, flatMap, hasParseErrors, invertChangeRange } from './utils';
+import { unixifyPath, hasSupportedExtension, addUnique, flatMap, hasParseErrors, invertChangeRange, djb2 } from './utils';
 import { Minimatch, IMinimatch } from 'minimatch';
 import { ProcessorLoader } from './services/processor-loader';
 import { injectable } from 'inversify';
@@ -118,8 +120,8 @@ export class Runner {
                 const originalContent = mapped === undefined ? sourceFile.text : mapped.originalContent;
                 let summary: FileSummary;
                 const fix = shouldFix(sourceFile, options, originalName);
-                // TODO consider reportUselessDirectives in cache
-                const resultFromCache = programState?.getUpToDateResult(sourceFile.fileName, effectiveConfig);
+                const configHash = programState === undefined ? undefined : createConfigHash(effectiveConfig, linterOptions);
+                const resultFromCache = programState?.getUpToDateResult(sourceFile.fileName, configHash!);
                 if (fix) {
                     let updatedFile = false;
                     summary = this.linter.lintAndFix(
@@ -164,7 +166,7 @@ export class Runner {
                     };
                 }
                 if (programState !== undefined && resultFromCache !== summary.findings)
-                    programState.setFileResult(file, effectiveConfig, summary.findings);
+                    programState.setFileResult(file, configHash!, summary.findings);
                 yield [originalName, summary];
             }
             programState?.save();
@@ -445,6 +447,31 @@ function shouldFix(sourceFile: ts.SourceFile, options: Pick<LintOptions, 'fix'>,
     }
     return options.fix;
 }
+
+
+function createConfigHash(config: ReducedConfiguration, linterOptions: LinterOptions) {
+    return '' + djb2(JSON.stringify({
+        rules: mapToObject(config.rules, stripRuleConfig),
+        settings: mapToObject(config.settings, identity),
+        reportUselessDirectives: linterOptions.reportUselessDirectives,
+    }));
+}
+
+function mapToObject<T, U>(map: ReadonlyMap<string, T>, transform: (v: T) => U) {
+    const result: Record<string, U> = {};
+    for (const [key, value] of map)
+        result[key] = transform(value);
+    return result;
+}
+
+function identity<T>(v: T) {
+    return v;
+}
+
+function stripRuleConfig({rulesDirectories: _ignored, ...rest}: EffectiveConfiguration.RuleConfig) {
+    return rest;
+}
+
 
 declare module 'typescript' {
     function matchFiles(
